UI: Hovering a player no shows a list of point events
In the UI, a user can now get a list of point events that shows the
source of a player's points. The list is only shown when a player's
name is hovered and disappears otherwise.

The current solution was a quick one and is not very clean, as the UI
may change soon anyway.