UI: The UI now gives an indication of whether or not a player's team has played
The UI now shows player name in green if they gameweek has started
but has not played (meaning they are still able to get points).

A future addition to this is to give an indication of if a player's
team is/has played but have no points (will be automatically benched).